requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
comm==0.2.3
config==0.5.1
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
executing==2.2.1
flask==3.1.2
flask-apscheduler==1.13.1
flask-cors==6.0.1
greenlet==3.2.4
groq==0.32.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
ipykernel==7.0.1
ipython==9.6.0
ipython-pygments-lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-client==8.6.3
jupyter-core==5.8.1
langchain==0.3.27
langchain-core==0.3.79
langchain-groq==0.3.8
langchain-mcp-adapters==0.1.11
langchain-text-splitters==0.3.11
langgraph==0.6.10
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.35
markupsafe==3.0.3
matplotlib-inline==0.1.7
mcp==1.17.0
nest-asyncio==1.6.0
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
parso==0.8.5
platformdirs==4.5.0
prompt-toolkit==3.0.52
psutil==7.1.0
psycopg2==2.9.11
pure-eval==0.2.3
pydantic==2.12.2
pydantic-core==2.41.4
pydantic-settings==2.11.0
pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pywin32==311
pyyaml==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.44
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.48.0
tenacity==9.1.2
tornado==6.5.2
traitlets==5.14.3
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
uvicorn==0.37.0
wcwidth==0.2.14
werkzeug==3.1.3
xxhash==3.6.0
zstandard==0.25.0
